refactor(cell/content): split chain_replay_tests by mission family (#244)
chain_replay_tests.rs was 919 lines (over the 700-line cap). Split by
mission family into a chain_replay_tests/ directory:

  chain_replay_tests/mod.rs          - file-level docs + mission sub-module declarations
  chain_replay_tests/mission_1562.rs - chain 3026 dialog acceptance gating (3 tests)
  chain_replay_tests/mission_639.rs  - chain 1034 ambernol vial remove_item (1 test)
  chain_replay_tests/mission_641.rs  - chain 1051 briefing re-fire prevention (2 tests)
  chain_replay_tests/mission_638.rs  - chains 1011/1012 archetype-gated dialog routing (4 tests + 2 helpers)
  chain_replay_tests/mission_687.rs  - chains 1097/1098/1099/1103 Aftermath (5 tests)

The two `assert_region_enter_*` helpers are only consumed by mission 638's
archetype-routing tests, so they live alongside them rather than in a shared
support module.

All 16 tests pass; no behavior change.
